backend/requirements.txt
fastapi[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pydantic
pydantic-settings
python-dotenv
asyncpg
greenlet
alembic
